+++ src/main/java/io/xpring/xrpl/DefaultXpringClient.java
@@ -87,6 +87,9 @@ public class DefaultXpringClient implements XpringClientDecorator {
     /**
      * Retrieve the transaction status for a given transaction hash.
      *
+     * Note: This API will return UNKNOWN for any transaction that is not a Payment transaction or has the
+     * tfPartialPayment flag set.
+     *
      * @param transactionHash The hash of the transaction.
      * @return The status of the given transaction.
      */
@@ -94,6 +97,10 @@ public class DefaultXpringClient implements XpringClientDecorator {
         Objects.requireNonNull(transactionHash);
 
         RawTransactionStatus transactionStatus = getRawTransactionStatus(transactionHash);
+
+        if (!transactionStatus.isFullPayment()) {
+            return TransactionStatus.UNKNOWN;
+        }
 
         // Return PENDING if the transaction is not validated.
         if (!transactionStatus.getValidated()) {